+++ packages/thumbprint-react/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+-   [Patch] Update `ConditionalPortal` to explicitly allow and handle null contents. (#357)
 -   [Patch] Update description of Avatar prop.
 
 ## 6.3.2 - 2019-07-01